Move resource mata data to resource package
