+++ shardingsphere-infra/shardingsphere-infra-common/src/test/java/org/apache/shardingsphere/infra/metadata/resource/CachedDatabaseMetaDataTest.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.infra.metadata.schema.model.datasource;
+package org.apache.shardingsphere.infra.metadata.resource;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
